Update plugins.md
I have made a small suggested change to the plugins tutorial that clarifies the different ways the plugins can be registered. It was something I was not clear on at first and had to figure out a little myself.

I think this change makes it clearer that the simplest way to register a plugin is to pass register "a plugin registration function". Then progresses to an array of registration functions then expands that you can pass an object with register and options properties and then finally an array of these objects.

It seems also that this was the original intention of the tutorial.

I have only just started developing with Hapi in the last few days so I apologise if I have misunderstood something about the intention here.

I also note that you can mix functions and objects in the array but this is probably an uanessacry complication to add to this tutorial so I have not added an example of this.